Get longer

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int [] sales = {3, 44, 3, 4, 5, 6};
+        long [] sales = {3, 44, 3, 4, 5, 6};
         SalesManager salesManager = new SalesManager(sales);
         salesManager.max();
         System.out.println(salesManager.max());
